Renaming Alert Model and AlertComponent
