+++ src/app/core/services/worker.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Alert } from '@core/model/alerts.model';
+import { Alert } from '@core/model/alert.model';
 import {
   AvailableQualificationsResponse,
   QualificationRequest,
